requirements.txt
streamlit==1.35.0
plotly==5.22.0
pandas==2.2.2
google-generativeai==0.7.2
python-dotenv==1.0.1
faker==25.2.0
numpy==1.26.4